+++ R/use_visc_report.R
@@ -89,7 +89,7 @@ use_project_objectives <- function(study_name) {
 
 use_group_colors <- function(study_name) {
   usethis::use_template(
-    template = "group_colors.R",
+    template = "visc-project-group-colors.R",
     save_as = "R/group_colors.R",
     data = list(study_name = study_name),
     package = "VISCtemplates"
@@ -98,7 +98,7 @@ use_group_colors <- function(study_name) {
 
 use_study_schema <- function(study_name) {
   usethis::use_template(
-    template = "study_schema.R",
+    template = "visc-project-study-schema.R",
     save_as = "R/study_schema.R",
     data = list(study_name = study_name),
     package = "VISCtemplates"
@@ -107,7 +107,7 @@ use_study_schema <- function(study_name) {
 
 use_bib <- function(study_name) {
   usethis::use_template(
-    template = "bibliography.bib",
+    template = "visc-project-bibliography.bib",
     save_as = "docs/bibliography.bib",
     data = list(study_name = study_name),
     package = "VISCtemplates"
@@ -134,12 +134,16 @@ use_bib <- function(study_name) {
 #'   report_type = "bama"
 #'   )
 #' }
-use_visc_report <- function(report_name = "PTreport",
+use_visc_report <- function(report_name = "VDCnnn_assay_PTreport",
                             path = ".",
                             report_type = c("empty", "generic", "bama", "nab", "adcc"),
                             interactive = TRUE) {
 
-  stopifnot(report_type %in% c("empty", "generic", "bama", "nab", "adcc"))
+  report_type <- match.arg(report_type)
+
+  if (dirname(report_name) != ".") {
+    stop("report_name cannot include a subdirectory. you can instead specify a subdirectory using the path argument of use_visc_report().")
+  }
 
   # suppress usethis output when non-interactive
   old_usethis_quiet <- getOption('usethis.quiet')
@@ -147,28 +151,75 @@ use_visc_report <- function(report_name = "PTreport",
   options(usethis.quiet = !interactive)
 
   if (report_type != 'empty') challenge_visc_report(report_name, interactive)
-  if (!dir.exists(path)) dir.create(path, recursive = TRUE)
-  use_template <- paste0(
+
+  # create assay level folder (specified in path) and readme, if don't yet exist
+  if (!dir.exists(path)) {
+    dir.create(path, recursive = TRUE)
+
+    usethis::use_template(
+      template = "README_assay_folder.md",
+      data = list(study_name = get_study_name(report_name), assay_name = get_assay_name(report_name)),
+      save_as = file.path(path, "README.md"),
+      package = "VISCtemplates"
+    )
+  }
+
+  # create report folder and main Rmd document
+  visc_report_type <- paste0(
     'visc', '_', if (report_type == 'empty') 'empty' else 'report'
   )
   rmarkdown::draft(
     file = file.path(path, report_name),
-    template = use_template,
-    package = "VISCtemplates",
+    template = system.file("templates", visc_report_type, package = "VISCtemplates"),
     edit = FALSE
   )
+
   usethis::ui_done(
-    glue::glue("Creating {{report_type}} VISC report at '{{file.path(path, report_name)}}'")
+    glue::glue("Created {{report_type}} VISC report at '{{file.path(path, report_name)}}'")
   )
-  if (report_type != 'empty') {
+
+  # create readme for report folder
+  usethis::use_template(
+    template = "README_report_folder.md",
+    data = list(),
+    save_as = file.path(path, report_name, "README.md"),
+    package = "VISCtemplates"
+  )
+
+  # add draft methods
+  if (report_type != 'empty'){
     use_visc_methods(path = file.path(path, report_name), assay = report_type,
                      interactive = interactive)
   }
 
+}
 
+# function to infer study name from report name
+get_study_name <- function(report_name) {
+  underscore_positions <- gregexpr('\\_', report_name)[[1]]
+  n_underscores <- length(underscore_positions[underscore_positions != -1])
+  if (n_underscores >= 2) {
+    study_name <- strsplit(report_name, '_')[[1]][1]
+  } else {
+    study_name <- "Study"
+  }
+  return(study_name)
+}
+
+# function to infer assay name from report name
+get_assay_name <- function(report_name) {
+  underscore_positions <- gregexpr('\\_', report_name)[[1]]
+  n_underscores <- length(underscore_positions[underscore_positions != -1])
+  if (n_underscores >= 2) {
+    assay_name <- strsplit(report_name, '_')[[1]][2]
+  } else {
+    assay_name <- "Assay"
+  }
+  return(assay_name)
 }
 
 challenge_visc_report <- function(report_name, interactive = TRUE) {
+
   if (!interactive) return(invisible(NULL))
 
   continue <- usethis::ui_yeah("